pyproject.toml
[project]
name = "receipt-scanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-genai>=1.9.0",
    "pillow>=11.1.0",
    "uvicorn>=0.34.0",
]
